*Major Update*
Major update in the core code of allee().
This is due to the meaning of terms Rn and Rm which were supposed to be
the growth rate but were in fact the product between the growth rate
and the population size in the previous step. This makes equating of
the theoretical maximum of the two strategies and the computation of
DELTA wrong.

The updated version moves the terms m[t-1] and n[t-1] outside the
definition of Rm and Rn, which are not pure growth rates. The
comparison is made on the basis of the growths rates, whilst the terms
m[t-1] and n[t-1] are reintroduced in the computation of the actual
number of individuals moving (C) and the computation of the final
population size.